tests/wpt/tests/html/semantics/embedded-content/the-img-element/update-src-complete.html

<!doctype html>
<meta charset="utf-8">
<title>Changing the img src should retain the 'complete' property</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<p id="display"><img src="image.png"></p>
<script>
    setup({ single_test: true });

    function check() {
        var img = document.querySelector("img");
        assert_true(img.complete, "By onload, image should have loaded");
        img.src = `image.png?${Math.random()}`;
        assert_false(img.complete, "Now that we're loading we should no longer be complete");
        img.onload = function () {
            assert_true(img.complete, "The new thing should have loaded.");
            done();
        }
    }

    onload = function () {
        check();
    };

</script>
